Создан модуль productsModule в store
